Make it possible to write chunks larger than 4 GB
Added a special chunk flag indicating that chunk size is a 64 bit value.
This allows to handle larger chunks (which heppens quite rarely) while
still maintaining backward compatibility with the majority of traces.

Bumped RDC file version of 0x101 (1.1) to make sure older version cannot
read the new file format.

+++ renderdoc/api/replay/structured_data.h
@@ -238,7 +238,7 @@ struct SDChunkMetaData
   DOCUMENT(R"(The length in bytes of this chunk - may be longer than the actual sum of the data if a
 conservative size estimate was used on creation to avoid seeking to fix-up the stored length.
 )");
-  uint32_t length = 0;
+  uint64_t length = 0;
 
   DOCUMENT("The ID of the thread where this chunk was recorded.");
   uint64_t threadID = 0;
