+++ renderdoc/driver/d3d11/d3d11_device.h
@@ -494,7 +494,7 @@ public:
   // log replaying
 
   bool Prepare_InitialState(ID3D11DeviceChild *res);
-  uint32_t GetSize_InitialState(ResourceId id, ID3D11DeviceChild *res);
+  uint64_t GetSize_InitialState(ResourceId id, ID3D11DeviceChild *res);
   template <typename SerialiserType>
   bool Serialise_InitialState(SerialiserType &ser, ResourceId resid, ID3D11DeviceChild *res);
