+++ renderdoc/driver/d3d11/d3d11_manager.h
@@ -262,7 +262,7 @@ private:
   bool Force_InitialState(ID3D11DeviceChild *res, bool prepare);
   bool Need_InitialStateChunk(ID3D11DeviceChild *res);
   bool Prepare_InitialState(ID3D11DeviceChild *res);
-  uint32_t GetSize_InitialState(ResourceId id, ID3D11DeviceChild *res);
+  uint64_t GetSize_InitialState(ResourceId id, ID3D11DeviceChild *res);
   bool Serialise_InitialState(WriteSerialiser &ser, ResourceId resid, ID3D11DeviceChild *res);
   void Create_InitialState(ResourceId id, ID3D11DeviceChild *live, bool hasData);
   void Apply_InitialState(ID3D11DeviceChild *live, D3D11InitialContents data);
